use warning banner for "you must be logged in to post" message
